Cross correlation function between two vectors

Calculate the lagged correlation between numeric vectors x and y. Vectors x and y are assumed to be captured at the same resolution and, similarly, successive values in x and y are assumed to be equi-distant. Missing values are allowed in each vector, correlations are calculated based on the complete cases.

Usage

get_ccf(x, y, min.overlap = round(0.1 * max(length(x), length(y))))

Arguments

x

vector, assumption is that x is longer than y

y

vector

min.overlap

integer value: what is the minimal number of values between x and y that should be considered?

Value

list with ccf values and lags

Details

This version of the cross correlation function is different from the stats implementation of ccf in two ways:

We consider the full region of correlations between vectors x and y as specified by min.overlap rather than just the overlap. The two vectors can be of very different length (e.g. when y is just a snippet recovered from a crime scene and x is from the full length object in the lab).

We do not use a Fourier transformation to calculate cross-correlation. This makes the evaluation slower, but prevents any edge effects.

Examples

library(dplyr)
x <- runif(20)
get_ccf(x, lead(x, 5))
#> $lag
#>  [1] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> [20]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#> 
#> $ccf
#>  [1]            NA            NA            NA            NA            NA
#>  [6]  1.0000000000  0.5606775495  0.5006372042 -0.8701294667 -0.2197378869
#> [11]  0.6638213282  0.4708427741 -0.3129834665 -0.1485806377  0.1567846418
#> [16]  0.2382472062 -0.1003719609  0.0752290966 -0.1153897806  0.5471846050
#> [21] -0.2239252608  0.2527398512 -0.1851625808  1.0000000000 -0.2539852150
#> [26]  0.3359820625 -0.1743128737  0.5903446869 -0.1602742353 -0.0459090454
#> [31]  0.0382289689  0.3321988283  0.0839392194 -0.0004708556 -0.5085078724
#> [36]  0.5469034532  1.0000000000
#> 
get_ccf(x, lag(x, 5), min.overlap = 3)
#> $lag
#>  [1] -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1
#> [20]   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#> 
#> $ccf
#>  [1]  0.66434096 -0.64361993 -0.37242775  0.13483731  0.36034385 -0.49391532
#>  [7]  0.32351181 -0.28436967  0.55077904 -0.51771634  0.13465736 -0.49778955
#> [13]  1.00000000 -0.34969879  0.04978736 -0.42967093  0.51329926 -0.11538978
#> [19]  0.07522910 -0.10037196  0.23824721  0.15678464 -0.14858064 -0.31298347
#> [25]  0.47084277  0.66382133 -0.21973789 -0.87012947  0.50063720  0.56067755
#> [31]          NA          NA          NA          NA          NA
#> 
x <- runif(100)
get_ccf(x[45:50], x, min.overlap = 6)
#> $lag
#>  [1] -94 -93 -92 -91 -90 -89 -88 -87 -86 -85 -84 -83 -82 -81 -80 -79 -78 -77 -76
#> [20] -75 -74 -73 -72 -71 -70 -69 -68 -67 -66 -65 -64 -63 -62 -61 -60 -59 -58 -57
#> [39] -56 -55 -54 -53 -52 -51 -50 -49 -48 -47 -46 -45 -44 -43 -42 -41 -40 -39 -38
#> [58] -37 -36 -35 -34 -33 -32 -31 -30 -29 -28 -27 -26 -25 -24 -23 -22 -21 -20 -19
#> [77] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> 
#> $ccf
#>  [1] -0.422241218  0.199760461  0.418717940 -0.035048566 -0.066696642
#>  [6]  0.139233093 -0.614492937 -0.233317441 -0.457972567 -0.234317679
#> [11]  0.516505003 -0.139913196  0.049364086  0.405296480 -0.056164503
#> [16] -0.639396471 -0.603234761  0.471519582  0.500047003 -0.041287906
#> [21]  0.434326714  0.779641800 -0.549315443 -0.501659692 -0.214535668
#> [26] -0.593983477 -0.471267920  0.433462547  0.229188014  0.375562920
#> [31]  0.085759368  0.577909331  0.495442386 -0.551936372 -0.522524238
#> [36]  0.193067861 -0.332623078 -0.725614124 -0.029579324  0.589586835
#> [41] -0.366082061  0.259331690  0.261414660 -0.331136819 -0.475859942
#> [46] -0.437397234  0.158655996  0.526928926 -0.074658840  0.436652825
#> [51]  1.000000000 -0.229996349 -0.359644613 -0.052916969 -0.431527683
#> [56] -0.148016015  0.003138872 -0.133486504  0.277461129  0.539335073
#> [61]  0.006569042 -0.486532713  0.237856050  0.078986096 -0.463796641
#> [66]  0.199367357  0.673207994  0.185337734 -0.419888168 -0.544153895
#> [71]  0.057951248  0.026950784 -0.448846098 -0.235976713  0.101836444
#> [76]  0.618104312  0.706119269 -0.664662955 -0.301735293  0.230730026
#> [81]  0.012706958 -0.385017685  0.128216731  0.682183705  0.362322521
#> [86] -0.034568892 -0.186250547 -0.130649787  0.156593670 -0.859865345
#> [91] -0.331872036  0.618403777 -0.700231713  0.139925377  0.526242478
#>