Extract number of matching striation marks from two aligned signatures

Extract number of matching striation marks from two aligned signatures

Usage

extract_feature_matches(striae)

Arguments

striae

data frame of striation marks based on two aligned signatures

Value

number of matching striation marks

See also

Examples

if (FALSE) {
# Set the data up to be read in, cleaned, etc.
library(bulletxtrctr)
library(x3ptools)

example_data <- bullet_pipeline(
  location = list(
    Bullet1 = c(hamby252demo$bullet1[2]),
    Bullet2 = c(hamby252demo$bullet2[4])
  ),
  x3p_clean = function(x) x %>%
      x3p_scale_unit(scale_by=10^6) %>%
      rotate_x3p(angle = -90) %>%
      y_flip_x3p()
)

alignment <- sig_align(example_data$sigs[[1]]$sig,
                       example_data$sigs[[2]]$sig)
striae <- sig_cms_max(alignment)
extract_feature_matches(striae$lines)
extract_feature_mismatches(striae$lines)
}