Cross correlation function between two vectors

Cross correlation function between two vectors

Usage

get_ccf(x, y, min.overlap = round(0.1 * max(length(x), length(y))))

Arguments

x

vector, assumption is that x is longer than y

y

vector

min.overlap

integer value: what is the minimal number of values between x and y that should be considered?

Value

list with ccf values and lags

Examples

library(dplyr)
x <- runif(20)
get_ccf(x, lead(x, 5))
#> $lag
#>  [1] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> [20]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#> 
#> $ccf
#>  [1]          NA          NA          NA          NA          NA  1.00000000
#>  [7] -0.23058980 -0.11880574  0.56234510 -0.37532997 -0.33973094  0.36787153
#> [13]  0.21161259 -0.29603278  0.21499366 -0.35950846 -0.39563049  0.24004493
#> [19] -0.02519986 -0.14507036  0.42176685  0.12193827 -0.01434766  1.00000000
#> [25] -0.21315774  0.07257053  0.69648426 -0.63562599  0.27985819  0.46967492
#> [31] -0.74050153  0.43466211  0.14852767 -0.80714272  0.89135729 -0.22302047
#> [37] -1.00000000
#> 
get_ccf(x, lag(x, 5), min.overlap = 3)
#> $lag
#>  [1] -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1
#> [20]   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#> 
#> $ccf
#>  [1]  0.83368573 -0.52369435 -0.02351379  0.18037292 -0.60977594 -0.40498938
#>  [7] -0.05655749 -0.08475632 -0.36242453  0.16639355  0.07537389 -0.06885116
#> [13]  1.00000000 -0.02467712  0.03252334  0.30513089 -0.22166044 -0.02519986
#> [19]  0.24004493 -0.39563049 -0.35950846  0.21499366 -0.29603278  0.21161259
#> [25]  0.36787153 -0.33973094 -0.37532997  0.56234510 -0.11880574 -0.23058980
#> [31]          NA          NA          NA          NA          NA
#> 
x <- runif(100)
get_ccf(x[45:50], x, min.overlap = 6)
#> $lag
#>  [1] -94 -93 -92 -91 -90 -89 -88 -87 -86 -85 -84 -83 -82 -81 -80 -79 -78 -77 -76
#> [20] -75 -74 -73 -72 -71 -70 -69 -68 -67 -66 -65 -64 -63 -62 -61 -60 -59 -58 -57
#> [39] -56 -55 -54 -53 -52 -51 -50 -49 -48 -47 -46 -45 -44 -43 -42 -41 -40 -39 -38
#> [58] -37 -36 -35 -34 -33 -32 -31 -30 -29 -28 -27 -26 -25 -24 -23 -22 -21 -20 -19
#> [77] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> 
#> $ccf
#>  [1] -0.300128813  0.453667563 -0.267096408  0.439575228 -0.525673620
#>  [6]  0.493332457  0.355356887  0.102857190 -0.255853117 -0.565387456
#> [11]  0.075459970  0.377840026 -0.227702803 -0.041472252  0.380448430
#> [16] -0.637663928  0.250442705  0.623340396 -0.396432696 -0.147749818
#> [21]  0.280236755 -0.560048217  0.360558112  0.255084688 -0.348559613
#> [26]  0.166016422  0.525337226 -0.046837880 -0.593711352 -0.065546948
#> [31]  0.206531872  0.132841904 -0.183512666  0.427789123 -0.779267816
#> [36]  0.094662166  0.548850165 -0.186156328 -0.137971120 -0.878998547
#> [41]  0.407944268  0.523102177  0.333277520 -0.433297193 -0.693286496
#> [46]  0.458539051  0.250754880  0.197448809 -0.633955768 -0.094810024
#> [51]  1.000000000 -0.035494859 -0.770559841 -0.616203353  0.290512524
#> [56]  0.671565346  0.223374743 -0.490285588 -0.566249582  0.510183735
#> [61]  0.848735659 -0.018788694 -0.807765903 -0.194758892 -0.184950937
#> [66] -0.194854656  0.168250018 -0.020557072  0.480212898  0.103523730
#> [71]  0.154108392  0.236081832 -0.863774559 -0.208290622 -0.216307864
#> [76] -0.453439687  0.122068758  0.540412563  0.447846217  0.060085931
#> [81] -0.125376036  0.010224719 -0.574225549  0.579873825  0.200753440
#> [86] -0.662978959  0.253754006 -0.565146836  0.382391128  0.576036709
#> [91] -0.006159097 -0.446549664 -0.862811187  0.120326043  0.646382399
#>