Cross correlation function between two vectors

Calculate the lagged correlation between numeric vectors x and y. Vectors x and y are assumed to be captured at the same resolution and, similarly, successive values in x and y are assumed to be equi-distant. Missing values are allowed in each vector, correlations are calculated based on the complete cases.

Usage

get_ccf(x, y, min.overlap = round(0.1 * max(length(x), length(y))))

Arguments

x

vector, assumption is that x is longer than y

y

vector

min.overlap

integer value: what is the minimal number of values between x and y that should be considered?

Value

list with ccf values and lags

Details

This version of the cross correlation function is different from the stats implementation of ccf in two ways:

We consider the full region of correlations between vectors x and y as specified by min.overlap rather than just the overlap. The two vectors can be of very different length (e.g. when y is just a snippet recovered from a crime scene and x is from the full length object in the lab).

We do not use a Fourier transformation to calculate cross-correlation. This makes the evaluation slower, but prevents any edge effects.

Examples

library(dplyr)
x <- runif(20)
get_ccf(x, lead(x, 5))
#> $lag
#>  [1] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> [20]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#> 
#> $ccf
#>  [1]           NA           NA           NA           NA           NA
#>  [6]  1.000000000  0.200413238  0.451209205  0.708597624 -0.104472470
#> [11] -0.221546855 -0.030330256 -0.484368827  0.041423443  0.255613287
#> [16]  0.324209327  0.299180432  0.091349182  0.113324250 -0.509120218
#> [21] -0.067440795 -0.084427998 -0.139369333  1.000000000 -0.178146657
#> [26]  0.007810524 -0.179752321 -0.478513293  0.166435142 -0.194812334
#> [31]  0.693016686 -0.008026434  0.095294526  0.065065380 -0.901132921
#> [36]  0.939972215 -1.000000000
#> 
get_ccf(x, lag(x, 5), min.overlap = 3)
#> $lag
#>  [1] -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1
#> [20]   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#> 
#> $ccf
#>  [1]  0.55190293  0.89377003  0.11913114  0.59875950  0.70976128 -0.11613528
#>  [7]  0.15395950 -0.05811084 -0.68423217  0.35832190  0.09287232  0.01910049
#> [13]  1.00000000 -0.02094733 -0.10507070  0.12266594 -0.51367964  0.11332425
#> [19]  0.09134918  0.29918043  0.32420933  0.25561329  0.04142344 -0.48436883
#> [25] -0.03033026 -0.22154685 -0.10447247  0.70859762  0.45120920  0.20041324
#> [31]          NA          NA          NA          NA          NA
#> 
x <- runif(100)
get_ccf(x[45:50], x, min.overlap = 6)
#> $lag
#>  [1] -94 -93 -92 -91 -90 -89 -88 -87 -86 -85 -84 -83 -82 -81 -80 -79 -78 -77 -76
#> [20] -75 -74 -73 -72 -71 -70 -69 -68 -67 -66 -65 -64 -63 -62 -61 -60 -59 -58 -57
#> [39] -56 -55 -54 -53 -52 -51 -50 -49 -48 -47 -46 -45 -44 -43 -42 -41 -40 -39 -38
#> [58] -37 -36 -35 -34 -33 -32 -31 -30 -29 -28 -27 -26 -25 -24 -23 -22 -21 -20 -19
#> [77] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> 
#> $ccf
#>  [1] -0.6219845822  0.0038295044  0.8223292831 -0.3454404151  0.0896233824
#>  [6] -0.5509626205  0.3451241574  0.0182663016 -0.5103282115  0.4447109101
#> [11]  0.4542573234  0.2215785367 -0.7021568300  0.1315079212 -0.3493232721
#> [16]  0.2819609459  0.6787836936 -0.4446946100 -0.5330812706 -0.0883443726
#> [21]  0.9513036437 -0.2985160631 -0.5078294199 -0.1015345190  0.6388384950
#> [26]  0.5889234148 -0.5708593904 -0.5134281479  0.3359466427  0.5076661301
#> [31]  0.0268510688  0.1707294384 -0.1578355970 -0.5810213736  0.6973348359
#> [36] -0.4115255993 -0.1572294551  0.5318792123  0.2670978352 -0.0027131076
#> [41]  0.7248805122  0.0804196120 -0.1337251766 -0.5718741188 -0.0222763068
#> [46] -0.3192403962 -0.3930254806  0.5458162385 -0.2363412762 -0.1727500888
#> [51]  1.0000000000 -0.1848832383 -0.3666467708  0.4794460858 -0.3845421047
#> [56]  0.4598786453 -0.0005137689 -0.6929363872 -0.3530204110  0.6712490056
#> [61]  0.3592617585 -0.5682536947  0.2255320799  0.7430035136  0.1573005906
#> [66] -0.3172262506 -0.4926729627  0.3957277441 -0.4927691925  0.4875870585
#> [71]  0.0383203734 -0.3961325409 -0.3856065537  0.9371589911 -0.1263562878
#> [76] -0.2109469356 -0.2924993769  0.1095506885  0.3054711226 -0.4932179862
#> [81]  0.4237999157  0.6981075086 -0.1260096792 -0.0071571938 -0.4841739446
#> [86] -0.3984358450 -0.0983030997  0.4133466719  0.4738740684 -0.8174550203
#> [91] -0.3845617631  0.1615358269  0.0177696648  0.6532697845  0.2847510906
#>