Cross correlation function between two vectors

Cross correlation function between two vectors

get_ccf(x, y, min.overlap = round(0.1 * max(length(x), length(y))))

Arguments

x
vector, assumption is that x is longer than y

y
vector

min.overlap
integer value: what is the minimal number of values between x and y that should be considered?

Value

list with ccf values and lags

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

x <- runif(20)
get_ccf(x, lead(x, 5))

#> $lag
#>  [1] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> [20]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#> 
#> $ccf
#>  [1]          NA          NA          NA          NA          NA  1.00000000
#>  [7]  0.54413707 -0.71703253 -0.44052957  0.07192240  0.32732305  0.05181610
#> [13]  0.21609531 -0.07362168  0.28102704  0.07014014 -0.21062721 -0.31863133
#> [19] -0.15597237 -0.14668910 -0.12146868 -0.21711398  0.23173342  1.00000000
#> [25]  0.26155115 -0.21059340 -0.20126545 -0.12402102  0.07930273 -0.06438724
#> [31] -0.69975002 -0.52400746  0.37708831  0.82507338  0.66525735 -0.87771334
#> [37]  1.00000000
#> 

get_ccf(x, lag(x, 5), min.overlap=3)

#> $lag
#>  [1] -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1
#> [20]   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#> 
#> $ccf
#>  [1]  0.867222169 -0.520883165 -0.684687565  0.312384545  0.602114415
#>  [6]  0.343532926 -0.498462183 -0.314647893  0.001236191 -0.306388616
#> [11] -0.376626548  0.191049485  1.000000000  0.189277573 -0.348056970
#> [16] -0.194645923  0.084268771 -0.155972372 -0.318631328 -0.210627209
#> [21]  0.070140143  0.281027037 -0.073621682  0.216095311  0.051816097
#> [26]  0.327323049  0.071922396 -0.440529573 -0.717032532  0.544137066
#> [31]           NA           NA           NA           NA           NA
#> 

x <- runif(100)
get_ccf(x[45:50], x, min.overlap=6)

#> $lag
#>  [1] -94 -93 -92 -91 -90 -89 -88 -87 -86 -85 -84 -83 -82 -81 -80 -79 -78 -77 -76
#> [20] -75 -74 -73 -72 -71 -70 -69 -68 -67 -66 -65 -64 -63 -62 -61 -60 -59 -58 -57
#> [39] -56 -55 -54 -53 -52 -51 -50 -49 -48 -47 -46 -45 -44 -43 -42 -41 -40 -39 -38
#> [58] -37 -36 -35 -34 -33 -32 -31 -30 -29 -28 -27 -26 -25 -24 -23 -22 -21 -20 -19
#> [77] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> 
#> $ccf
#>  [1] -0.1483180552 -0.4512230165 -0.0418162852  0.7892567084 -0.6026100066
#>  [6] -0.0001144979 -0.2545209392  0.6767377026 -0.5840893856  0.2397440134
#> [11] -0.6658088797  0.3525540389  0.0661448286 -0.2288165017 -0.5191924603
#> [16]  0.1337129742  0.4216791090  0.4815311794 -0.9008191980  0.0538105792
#> [21]  0.3117397722 -0.0715084704 -0.3320963209  0.2861469872  0.5296610875
#> [26] -0.7706514565  0.5598357479 -0.5611328513  0.5130579827 -0.3617271948
#> [31] -0.1892396226  0.3418766601  0.3373982200  0.0920535333 -0.7479286612
#> [36]  0.0996858952  0.4137062164 -0.3717255488  0.4195386003  0.0736898111
#> [41]  0.3653895495 -0.7155537559 -0.3402895096  0.5379283670 -0.2007022752
#> [46] -0.1962784240  0.1803876754  0.0037630901  0.2078490666 -0.4078679497
#> [51]  1.0000000000 -0.5101564110  0.2755755409 -0.6044883152  0.6315691872
#> [56] -0.8614419856  0.5382853078  0.2115774536 -0.5540420087  0.2718637559
#> [61] -0.5078583393  0.6060496484 -0.2704621543  0.5957083277 -0.5444331464
#> [66] -0.4624861787  0.0972969093 -0.1956333870  0.2616740873  0.4853235091
#> [71]  0.1058328314  0.3834706336 -0.7928610681 -0.1311920478  0.0165618956
#> [76]  0.0679607105  0.3631131225 -0.0294602169 -0.6729584764  0.1528610147
#> [81]  0.4580673709  0.2348210568 -0.5605358222  0.2021717858 -0.0518461503
#> [86]  0.1564169425 -0.7783453138  0.1328738051  0.2349666167  0.6826462161
#> [91] -0.6546979054 -0.2855948135  0.6571323319 -0.1408475685 -0.0633743743
#>