Cross correlation function between two vectors

Calculate the lagged correlation between numeric vectors x and y. Vectors x and y are assumed to be captured at the same resolution and, similarly, successive values in x and y are assumed to be equi-distant. Missing values are allowed in each vector, correlations are calculated based on the complete cases.

Usage

get_ccf(x, y, min.overlap = round(0.1 * max(length(x), length(y))))

Arguments

x

vector, assumption is that x is longer than y

y

vector

min.overlap

integer value: what is the minimal number of values between x and y that should be considered?

Value

list with ccf values and lags

Details

This version of the cross correlation function is different from the stats implementation of ccf in two ways:

We consider the full region of correlations between vectors x and y as specified by min.overlap rather than just the overlap. The two vectors can be of very different length (e.g. when y is just a snippet recovered from a crime scene and x is from the full length object in the lab).

We do not use a Fourier transformation to calculate cross-correlation. This makes the evaluation slower, but prevents any edge effects.

Examples

library(dplyr)
x <- runif(20)
get_ccf(x, lead(x, 5))
#> $lag
#>  [1] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> [20]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#> 
#> $ccf
#>  [1]          NA          NA          NA          NA          NA -1.00000000
#>  [7] -0.80154328  0.18195514 -0.84425534  0.27524958  0.49278427  0.03839042
#> [13] -0.19796369  0.22442087 -0.39416486  0.05020158 -0.07984086  0.44166346
#> [19] -0.07509714  0.02290286 -0.15584208 -0.03433239 -0.28082229  1.00000000
#> [25] -0.15308885 -0.10470508  0.10738215 -0.16072750 -0.31071100  0.61902816
#> [31] -0.31909829  0.30656748 -0.04631530 -0.63711385 -0.89190056  0.03707042
#> [37] -1.00000000
#> 
get_ccf(x, lag(x, 5), min.overlap = 3)
#> $lag
#>  [1] -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1
#> [20]   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#> 
#> $ccf
#>  [1] -0.120358896 -0.387790752  0.902594783 -0.692011411  0.136707592
#>  [6] -0.134215348  0.371055532  0.118143529 -0.018421527 -0.316831029
#> [11]  0.032171444 -0.433230411  1.000000000 -0.317376205 -0.053640095
#> [16] -0.131042692 -0.001511274 -0.075097136  0.441663461 -0.079840859
#> [21]  0.050201576 -0.394164864  0.224420866 -0.197963695  0.038390423
#> [26]  0.492784269  0.275249580 -0.844255341  0.181955143 -0.801543276
#> [31]           NA           NA           NA           NA           NA
#> 
x <- runif(100)
get_ccf(x[45:50], x, min.overlap = 6)
#> $lag
#>  [1] -94 -93 -92 -91 -90 -89 -88 -87 -86 -85 -84 -83 -82 -81 -80 -79 -78 -77 -76
#> [20] -75 -74 -73 -72 -71 -70 -69 -68 -67 -66 -65 -64 -63 -62 -61 -60 -59 -58 -57
#> [39] -56 -55 -54 -53 -52 -51 -50 -49 -48 -47 -46 -45 -44 -43 -42 -41 -40 -39 -38
#> [58] -37 -36 -35 -34 -33 -32 -31 -30 -29 -28 -27 -26 -25 -24 -23 -22 -21 -20 -19
#> [77] -18 -17 -16 -15 -14 -13 -12 -11 -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0
#> 
#> $ccf
#>  [1] -0.007644500  0.009618919  0.099492100  0.017070215 -0.124464484
#>  [6] -0.338006748 -0.694426349  0.369679321  0.202975910 -0.132818631
#> [11]  0.210463482 -0.834972562 -0.545285486  0.378981632  0.363848216
#> [16]  0.174054129  0.359034237 -0.182633520 -0.367628159  0.579655306
#> [21] -0.129361127 -0.072941357  0.269080535 -0.735748288  0.047440745
#> [26]  0.303833241 -0.345629359 -0.040349928 -0.103836472 -0.666340358
#> [31] -0.507055937  0.600246995  0.321645353 -0.301904257  0.466655074
#> [36]  0.390672826  0.161590929  0.911433689  0.007286646 -0.789789214
#> [41] -0.091917978  0.035106257  0.226284925  0.739572551  0.024136080
#> [46] -0.878648687 -0.918555758 -0.507847515 -0.004224639  0.646270476
#> [51]  1.000000000  0.703122883 -0.211280277 -0.474959151 -0.253526386
#> [56] -0.464947661 -0.278448466  0.584282538  0.241830489  0.146100865
#> [61]  0.653665404  0.344872024 -0.679921079 -0.346987030  0.045938234
#> [66] -0.042098079 -0.497808157 -0.314003979  0.259683346 -0.150886553
#> [71]  0.334046720  0.572822820 -0.075567153  0.219359841 -0.233042866
#> [76] -0.339403323 -0.136103724 -0.837199285 -0.413664415  0.311007251
#> [81] -0.244313978  0.225090366  0.540359517 -0.266022422 -0.303753551
#> [86] -0.065713911 -0.117294419 -0.048420098  0.012766855  0.431933613
#> [91]  0.289163816  0.488260590  0.540770944  0.182209744  0.316964753
#>